requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6

langchain>=0.3.3
langchain-core>=0.3.9
langchain-community>=0.3.2
langchain-openai>=0.2.2
langchain-text-splitters>=0.3.0

faiss-cpu>=1.8.0.post1

pypdf>=4.3.1
docx2txt>=0.8

python-dotenv>=1.0.1
pydantic>=2.9.2
pydantic-settings>=2.6.1

requests>=2.32.3

streamlit>=1.38.0
